Move serialization to a dedicated package
