+++ python/tests/plugins/first_party/test_help_dump_impl.py
@@ -3,9 +3,9 @@ import json
 from unittest import TestCase
 from collections import OrderedDict
 from aac.cli.aac_command import AacCommand, AacCommandArgument
-from aac.cli.aac_command_encoder import AacCommandEncoder, AacCommandArgumentEncoder
 from aac.plugins.first_party.help_dump.help_dump_impl import help_dump
 from aac.plugins.plugin_manager import get_plugin_commands
+from aac.serialization import AacCommandEncoder, AacCommandArgumentEncoder
 
 from tests.helpers.assertion import assert_plugin_success
 
@@ -18,8 +18,7 @@ class TestHelpDump(TestCase):
         command_name = "command-name"
         command_description = "A command description."
         command_arguments = [
-            AacCommandArgument(f"arg{i}", f"The arg{i} description.", data_type="str")
-            for i in range(default_num_args)
+            AacCommandArgument(f"arg{i}", f"The arg{i} description.", data_type="str") for i in range(default_num_args)
         ]
 
         for i in range(3):
@@ -35,7 +34,7 @@ class TestHelpDump(TestCase):
         for command in get_plugin_commands():
             self.assertIn(
                 json.dumps(self.expected_formatted_output(command.name, command.description, command.arguments)),
-                result.get_messages_as_string()
+                result.get_messages_as_string(),
             )
 
     def test_help_dump_json_dump_sort(self):
@@ -51,5 +50,5 @@ class TestHelpDump(TestCase):
         return {
             "name": name,
             "description": description,
-            "arguments": [AacCommandArgumentEncoder().default(arg) for arg in arguments]
+            "arguments": [AacCommandArgumentEncoder().default(arg) for arg in arguments],
         }
